requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
autopep8==1.6.0
boto3==1.21.0
botocore==1.24.0
certifi==2021.10.8
cycler==0.11.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.4.2
fonttools==4.29.1
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
jmespath==0.10.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.4.1
promise==2.3
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==1.7.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
Rx==1.6.1
s3transfer==0.5.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
whitenoise==5.3.0
